+++ apps/pleiepenger-sykt-barn/src/app/søknad/arbeidstid-step/components/ArbeidIPeriodeSpørsmål.tsx
@@ -96,6 +96,7 @@ const ArbeidIPeriodeSpørsmål: React.FunctionComponent<Props> = ({
                                     name={getFieldName(ArbeidIPeriodeFormField.erLiktHverUke)}
                                     legend={spørsmål.erLiktHverUke}
                                     validate={getArbeidIPeriodeErLiktHverUkeValidator(intlValues)}
+                                    renderHorizontal={false}
                                     labels={{
                                         yes: text(`arbeidIPeriode.erLiktHverUke.ja`),
                                         no: text(`arbeidIPeriode.erLiktHverUke.nei`),